Excel Homework
Answers to questions:

We can conclude that theater productions are the most successful kickstarter productions, and that plays are by far the most popular among kickstarter categories. Also, overall successes seemed to peak in the month of May.

This dataset is not taking into account any kickstarter projects that did not fit into the categories that were listed.

We could create a pie chart that would easily show what the proportions of successes that each category had. We could create a scatterplot to see if success was correlated with the amount of money spent on the project.